WIP: refactoring listen
