+++ src/features/listen/renderer/listenCapture.js
@@ -1,20 +1,29 @@
-// renderer.js
 const { ipcRenderer } = require('electron');
-const { makeStreamingChatCompletionWithPortkey } = require('../../common/services/aiProviderService.js');
 
-let mediaStream = null;
-let screenshotInterval = null;
-let audioContext = null;
-let audioProcessor = null;
-let micMediaStream = null;
-let audioBuffer = [];
+// ---------------------------
+// Constants & Globals
+// ---------------------------
 const SAMPLE_RATE = 24000;
 const AUDIO_CHUNK_DURATION = 0.1;
 const BUFFER_SIZE = 4096;
 
+const isLinux = process.platform === 'linux';
+const isMacOS = process.platform === 'darwin';
+
+let mediaStream = null;
+let micMediaStream = null;
+let screenshotInterval = null;
+let audioContext = null;
+let audioProcessor = null;
+let currentImageQuality = 'medium';
+let lastScreenshotBase64 = null;
+
 let systemAudioBuffer = [];
 const MAX_SYSTEM_BUFFER_SIZE = 10;
 
+// ---------------------------
+// Utility helpers (exact from renderer.js)
+// ---------------------------
 function isVoiceActive(audioFloat32Array, threshold = 0.005) {
     if (!audioFloat32Array || audioFloat32Array.length === 0) {
         return false;
@@ -30,11 +39,6 @@ function isVoiceActive(audioFloat32Array, threshold = 0.005) {
 
     return rms > threshold;
 }
-
-let currentImageQuality = 'medium'; // Store current image quality for manual screenshots
-let lastScreenshotBase64 = null; // Store the latest screenshot
-
-let realtimeConversationHistory = [];
 
 function base64ToFloat32Array(base64) {
     const binaryString = atob(base64);
@@ -54,11 +58,29 @@ function base64ToFloat32Array(base64) {
     return float32Array;
 }
 
-async function queryLoginState() {
-    const userState = await ipcRenderer.invoke('get-current-user');
-    return userState;
+function convertFloat32ToInt16(float32Array) {
+    const int16Array = new Int16Array(float32Array.length);
+    for (let i = 0; i < float32Array.length; i++) {
+        // Improved scaling to prevent clipping
+        const s = Math.max(-1, Math.min(1, float32Array[i]));
+        int16Array[i] = s < 0 ? s * 0x8000 : s * 0x7fff;
+    }
+    return int16Array;
 }
 
+function arrayBufferToBase64(buffer) {
+    let binary = '';
+    const bytes = new Uint8Array(buffer);
+    const len = bytes.byteLength;
+    for (let i = 0; i < len; i++) {
+        binary += String.fromCharCode(bytes[i]);
+    }
+    return btoa(binary);
+}
+
+// ---------------------------
+// Complete SimpleAEC implementation (exact from renderer.js)
+// ---------------------------
 class SimpleAEC {
     constructor() {
         this.adaptiveFilter = new Float32Array(1024);
@@ -179,11 +201,24 @@ class SimpleAEC {
 
 let aecProcessor = new SimpleAEC();
 
-const isLinux = process.platform === 'linux';
-const isMacOS = process.platform === 'darwin';
+// System audio data handler
+ipcRenderer.on('system-audio-data', (event, { data }) => {
+    systemAudioBuffer.push({
+        data: data,
+        timestamp: Date.now(),
+    });
 
-window.pickleGlass = window.pickleGlass || {};
+    // 오래된 데이터 제거
+    if (systemAudioBuffer.length > MAX_SYSTEM_BUFFER_SIZE) {
+        systemAudioBuffer = systemAudioBuffer.slice(-MAX_SYSTEM_BUFFER_SIZE);
+    }
 
+    console.log('📥 Received system audio for AEC reference');
+});
+
+// ---------------------------
+// Complete token tracker (exact from renderer.js)
+// ---------------------------
 let tokenTracker = {
     tokens: [],
     audioStartTime: null,
@@ -265,126 +300,201 @@ setInterval(() => {
     tokenTracker.trackAudioTokens();
 }, 2000);
 
-function pickleGlassElement() {
-    return document.getElementById('pickle-glass');
-}
+// ---------------------------
+// Audio processing functions (exact from renderer.js)
+// ---------------------------
+function setupMicProcessing(micStream) {
+    const micAudioContext = new AudioContext({ sampleRate: SAMPLE_RATE });
+    const micSource = micAudioContext.createMediaStreamSource(micStream);
+    const micProcessor = micAudioContext.createScriptProcessor(BUFFER_SIZE, 1, 1);
 
-function convertFloat32ToInt16(float32Array) {
-    const int16Array = new Int16Array(float32Array.length);
-    for (let i = 0; i < float32Array.length; i++) {
-        // Improved scaling to prevent clipping
-        const s = Math.max(-1, Math.min(1, float32Array[i]));
-        int16Array[i] = s < 0 ? s * 0x8000 : s * 0x7fff;
-    }
-    return int16Array;
-}
+    let audioBuffer = [];
+    const samplesPerChunk = SAMPLE_RATE * AUDIO_CHUNK_DURATION;
 
-function arrayBufferToBase64(buffer) {
-    let binary = '';
-    const bytes = new Uint8Array(buffer);
-    const len = bytes.byteLength;
-    for (let i = 0; i < len; i++) {
-        binary += String.fromCharCode(bytes[i]);
-    }
-    return btoa(binary);
-}
+    micProcessor.onaudioprocess = async e => {
+        const inputData = e.inputBuffer.getChannelData(0);
+        audioBuffer.push(...inputData);
 
-async function initializeopenai(profile = 'interview', language = 'en') {
-    // The API key is now handled in the main process from .env file.
-    // We just need to trigger the initialization.
-    try {
-        console.log(`Requesting OpenAI initialization with profile: ${profile}, language: ${language}`);
-        const success = await ipcRenderer.invoke('initialize-openai', profile, language);
-        if (success) {
-            // The status will be updated via 'update-status' event from the main process.
-            console.log('OpenAI initialization successful.');
-        } else {
-            console.error('OpenAI initialization failed.');
-            const appElement = pickleGlassElement();
-            if (appElement && typeof appElement.setStatus === 'function') {
-                appElement.setStatus('Initialization Failed');
+        while (audioBuffer.length >= samplesPerChunk) {
+            let chunk = audioBuffer.splice(0, samplesPerChunk);
+            let processedChunk = new Float32Array(chunk);
+
+            // Check for system audio and apply AEC only if voice is active
+            if (aecProcessor && systemAudioBuffer.length > 0) {
+                const latestSystemAudio = systemAudioBuffer[systemAudioBuffer.length - 1];
+                const systemFloat32 = base64ToFloat32Array(latestSystemAudio.data);
+
+                // Apply AEC only when system audio has active speech
+                if (isVoiceActive(systemFloat32)) {
+                    processedChunk = aecProcessor.process(new Float32Array(chunk), systemFloat32);
+                    console.log('🔊 Applied AEC because system audio is active');
+                }
             }
+
+            const pcmData16 = convertFloat32ToInt16(processedChunk);
+            const base64Data = arrayBufferToBase64(pcmData16.buffer);
+
+            await ipcRenderer.invoke('send-audio-content', {
+                data: base64Data,
+                mimeType: 'audio/pcm;rate=24000',
+            });
+        }
+    };
+
+    micSource.connect(micProcessor);
+    micProcessor.connect(micAudioContext.destination);
+
+    audioProcessor = micProcessor;
+}
+
+function setupLinuxMicProcessing(micStream) {
+    // Setup microphone audio processing for Linux
+    const micAudioContext = new AudioContext({ sampleRate: SAMPLE_RATE });
+    const micSource = micAudioContext.createMediaStreamSource(micStream);
+    const micProcessor = micAudioContext.createScriptProcessor(BUFFER_SIZE, 1, 1);
+
+    let audioBuffer = [];
+    const samplesPerChunk = SAMPLE_RATE * AUDIO_CHUNK_DURATION;
+
+    micProcessor.onaudioprocess = async e => {
+        const inputData = e.inputBuffer.getChannelData(0);
+        audioBuffer.push(...inputData);
+
+        // Process audio in chunks
+        while (audioBuffer.length >= samplesPerChunk) {
+            const chunk = audioBuffer.splice(0, samplesPerChunk);
+            const pcmData16 = convertFloat32ToInt16(chunk);
+            const base64Data = arrayBufferToBase64(pcmData16.buffer);
+
+            await ipcRenderer.invoke('send-audio-content', {
+                data: base64Data,
+                mimeType: 'audio/pcm;rate=24000',
+            });
+        }
+    };
+
+    micSource.connect(micProcessor);
+    micProcessor.connect(micAudioContext.destination);
+
+    // Store processor reference for cleanup
+    audioProcessor = micProcessor;
+}
+
+function setupWindowsLoopbackProcessing() {
+    // Setup audio processing for Windows loopback audio only
+    audioContext = new AudioContext({ sampleRate: SAMPLE_RATE });
+    const source = audioContext.createMediaStreamSource(mediaStream);
+    audioProcessor = audioContext.createScriptProcessor(BUFFER_SIZE, 1, 1);
+
+    let audioBuffer = [];
+    const samplesPerChunk = SAMPLE_RATE * AUDIO_CHUNK_DURATION;
+
+    audioProcessor.onaudioprocess = async e => {
+        const inputData = e.inputBuffer.getChannelData(0);
+        audioBuffer.push(...inputData);
+
+        // Process audio in chunks
+        while (audioBuffer.length >= samplesPerChunk) {
+            const chunk = audioBuffer.splice(0, samplesPerChunk);
+            const pcmData16 = convertFloat32ToInt16(chunk);
+            const base64Data = arrayBufferToBase64(pcmData16.buffer);
+
+            await ipcRenderer.invoke('send-audio-content', {
+                data: base64Data,
+                mimeType: 'audio/pcm;rate=24000',
+            });
+        }
+    };
+
+    source.connect(audioProcessor);
+    audioProcessor.connect(audioContext.destination);
+}
+
+// ---------------------------
+// Screenshot functions (exact from renderer.js)
+// ---------------------------
+async function captureScreenshot(imageQuality = 'medium', isManual = false) {
+    console.log(`Capturing ${isManual ? 'manual' : 'automated'} screenshot...`);
+
+    // Check rate limiting for automated screenshots only
+    if (!isManual && tokenTracker.shouldThrottle()) {
+        console.log('⚠️ Automated screenshot skipped due to rate limiting');
+        return;
+    }
+
+    try {
+        // Request screenshot from main process
+        const result = await ipcRenderer.invoke('capture-screenshot', {
+            quality: imageQuality,
+        });
+
+        if (result.success && result.base64) {
+            // Store the latest screenshot
+            lastScreenshotBase64 = result.base64;
+
+            // Note: sendResult is not defined in the original, this was likely an error
+            // Commenting out this section as it references undefined variable
+            /*
+            if (sendResult.success) {
+                // Track image tokens after successful send
+                const imageTokens = tokenTracker.calculateImageTokens(result.width || 1920, result.height || 1080);
+                tokenTracker.addTokens(imageTokens, 'image');
+                console.log(`📊 Image sent successfully - ${imageTokens} tokens used (${result.width}x${result.height})`);
+            } else {
+                console.error('Failed to send image:', sendResult.error);
+            }
+            */
+        } else {
+            console.error('Failed to capture screenshot:', result.error);
         }
     } catch (error) {
-        console.error('Error during OpenAI initialization IPC call:', error);
-        const appElement = pickleGlassElement();
-        if (appElement && typeof appElement.setStatus === 'function') {
-            appElement.setStatus('Error');
-        }
+        console.error('Error capturing screenshot:', error);
     }
 }
 
+async function captureManualScreenshot(imageQuality = null) {
+    console.log('Manual screenshot triggered');
+    const quality = imageQuality || currentImageQuality;
+    await captureScreenshot(quality, true);
+}
 
-ipcRenderer.on('system-audio-data', (event, { data }) => {
-    systemAudioBuffer.push({
-        data: data,
-        timestamp: Date.now(),
-    });
+async function getCurrentScreenshot() {
+    try {
+        // First try to get a fresh screenshot from main process
+        const result = await ipcRenderer.invoke('get-current-screenshot');
 
-    // 오래된 데이터 제거
-    if (systemAudioBuffer.length > MAX_SYSTEM_BUFFER_SIZE) {
-        systemAudioBuffer = systemAudioBuffer.slice(-MAX_SYSTEM_BUFFER_SIZE);
-    }
-
-    console.log('📥 Received system audio for AEC reference');
-});
-
-// Listen for status updates
-ipcRenderer.on('update-status', (event, status) => {
-    console.log('Status update:', status);
-    pickleGlass.e().setStatus(status);
-});
-
-// Listen for real-time STT updates
-ipcRenderer.on('stt-update', (event, data) => {
-    console.log('Renderer.js stt-update', data);
-    const { speaker, text, isFinal, isPartial, timestamp } = data;
-
-    if (isPartial) {
-        console.log(`🔄 [${speaker} - partial]: ${text}`);
-    } else if (isFinal) {
-        console.log(`✅ [${speaker} - final]: ${text}`);
-
-        const speakerText = speaker.toLowerCase();
-        const conversationText = `${speakerText}: ${text.trim()}`;
-
-        realtimeConversationHistory.push(conversationText);
-
-        if (realtimeConversationHistory.length > 30) {
-            realtimeConversationHistory = realtimeConversationHistory.slice(-30);
+        if (result.success && result.base64) {
+            console.log('📸 Got fresh screenshot from main process');
+            return result.base64;
         }
 
-        console.log(`📝 Updated realtime conversation history: ${realtimeConversationHistory.length} texts`);
-        console.log(`📋 Latest text: ${conversationText}`);
-    }
-
-    if (pickleGlass.e() && typeof pickleGlass.e().updateRealtimeTranscription === 'function') {
-        pickleGlass.e().updateRealtimeTranscription({
-            speaker,
-            text,
-            isFinal,
-            isPartial,
-            timestamp,
+        // If no screenshot available, capture one now
+        console.log('📸 No screenshot available, capturing new one');
+        const captureResult = await ipcRenderer.invoke('capture-screenshot', {
+            quality: currentImageQuality,
         });
+
+        if (captureResult.success && captureResult.base64) {
+            lastScreenshotBase64 = captureResult.base64;
+            return captureResult.base64;
+        }
+
+        // Fallback to last stored screenshot
+        if (lastScreenshotBase64) {
+            console.log('📸 Using cached screenshot');
+            return lastScreenshotBase64;
+        }
+
+        throw new Error('Failed to get screenshot');
+    } catch (error) {
+        console.error('Error getting current screenshot:', error);
+        return null;
     }
-});
+}
 
-
-ipcRenderer.on('update-structured-data', (_, structuredData) => {
-    console.log('📥 Received structured data update:', structuredData);
-    window.pickleGlass.structuredData = structuredData;
-    window.pickleGlass.setStructuredData(structuredData);
-});
-window.pickleGlass.structuredData = {
-    summary: [],
-    topic: { header: '', bullets: [] },
-    actions: [],
-};
-window.pickleGlass.setStructuredData = data => {
-    window.pickleGlass.structuredData = data;
-    pickleGlass.e()?.updateStructuredData?.(data);
-};
-
+// ---------------------------
+// Main capture functions (exact from renderer.js)
+// ---------------------------
 async function startCapture(screenshotIntervalSeconds = 5, imageQuality = 'medium') {
     // Store the image quality for manual screenshots
     currentImageQuality = imageQuality;
@@ -490,12 +600,6 @@ async function startCapture(screenshotIntervalSeconds = 5, imageQuality = 'mediu
             setupWindowsLoopbackProcessing();
         }
 
-        // console.log('MediaStream obtained:', {
-        //     hasVideo: mediaStream.getVideoTracks().length > 0,
-        //     hasAudio: mediaStream.getAudioTracks().length > 0,
-        //     videoTrack: mediaStream.getVideoTracks()[0]?.getSettings(),
-        // });
-
         // Start capturing screenshots - check if manual mode
         if (screenshotIntervalSeconds === 'manual' || screenshotIntervalSeconds === 'Manual') {
             console.log('Manual mode enabled - screenshots will be captured on demand only');
@@ -511,161 +615,10 @@ async function startCapture(screenshotIntervalSeconds = 5, imageQuality = 'mediu
         }
     } catch (err) {
         console.error('Error starting capture:', err);
-        pickleGlass.e().setStatus('error');
+        // Note: pickleGlass.e() is not available in this context, commenting out
+        // pickleGlass.e().setStatus('error');
     }
 }
-
-function setupMicProcessing(micStream) {
-    const micAudioContext = new AudioContext({ sampleRate: SAMPLE_RATE });
-    const micSource = micAudioContext.createMediaStreamSource(micStream);
-    const micProcessor = micAudioContext.createScriptProcessor(BUFFER_SIZE, 1, 1);
-
-    let audioBuffer = [];
-    const samplesPerChunk = SAMPLE_RATE * AUDIO_CHUNK_DURATION;
-
-    micProcessor.onaudioprocess = async e => {
-        const inputData = e.inputBuffer.getChannelData(0);
-        audioBuffer.push(...inputData);
-
-        while (audioBuffer.length >= samplesPerChunk) {
-            let chunk = audioBuffer.splice(0, samplesPerChunk);
-            let processedChunk = new Float32Array(chunk);
-
-            // Check for system audio and apply AEC only if voice is active
-            if (aecProcessor && systemAudioBuffer.length > 0) {
-                const latestSystemAudio = systemAudioBuffer[systemAudioBuffer.length - 1];
-                const systemFloat32 = base64ToFloat32Array(latestSystemAudio.data);
-
-                // Apply AEC only when system audio has active speech
-                if (isVoiceActive(systemFloat32)) {
-                    processedChunk = aecProcessor.process(new Float32Array(chunk), systemFloat32);
-                    console.log('🔊 Applied AEC because system audio is active');
-                }
-            }
-
-            const pcmData16 = convertFloat32ToInt16(processedChunk);
-            const base64Data = arrayBufferToBase64(pcmData16.buffer);
-
-            await ipcRenderer.invoke('send-audio-content', {
-                data: base64Data,
-                mimeType: 'audio/pcm;rate=24000',
-            });
-        }
-    };
-
-    micSource.connect(micProcessor);
-    micProcessor.connect(micAudioContext.destination);
-
-    audioProcessor = micProcessor;
-}
-////////// for index & subjects //////////
-
-function setupLinuxMicProcessing(micStream) {
-    // Setup microphone audio processing for Linux
-    const micAudioContext = new AudioContext({ sampleRate: SAMPLE_RATE });
-    const micSource = micAudioContext.createMediaStreamSource(micStream);
-    const micProcessor = micAudioContext.createScriptProcessor(BUFFER_SIZE, 1, 1);
-
-    let audioBuffer = [];
-    const samplesPerChunk = SAMPLE_RATE * AUDIO_CHUNK_DURATION;
-
-    micProcessor.onaudioprocess = async e => {
-        const inputData = e.inputBuffer.getChannelData(0);
-        audioBuffer.push(...inputData);
-
-        // Process audio in chunks
-        while (audioBuffer.length >= samplesPerChunk) {
-            const chunk = audioBuffer.splice(0, samplesPerChunk);
-            const pcmData16 = convertFloat32ToInt16(chunk);
-            const base64Data = arrayBufferToBase64(pcmData16.buffer);
-
-            await ipcRenderer.invoke('send-audio-content', {
-                data: base64Data,
-                mimeType: 'audio/pcm;rate=24000',
-            });
-        }
-    };
-
-    micSource.connect(micProcessor);
-    micProcessor.connect(micAudioContext.destination);
-
-    // Store processor reference for cleanup
-    audioProcessor = micProcessor;
-}
-
-function setupWindowsLoopbackProcessing() {
-    // Setup audio processing for Windows loopback audio only
-    audioContext = new AudioContext({ sampleRate: SAMPLE_RATE });
-    const source = audioContext.createMediaStreamSource(mediaStream);
-    audioProcessor = audioContext.createScriptProcessor(BUFFER_SIZE, 1, 1);
-
-    let audioBuffer = [];
-    const samplesPerChunk = SAMPLE_RATE * AUDIO_CHUNK_DURATION;
-
-    audioProcessor.onaudioprocess = async e => {
-        const inputData = e.inputBuffer.getChannelData(0);
-        audioBuffer.push(...inputData);
-
-        // Process audio in chunks
-        while (audioBuffer.length >= samplesPerChunk) {
-            const chunk = audioBuffer.splice(0, samplesPerChunk);
-            const pcmData16 = convertFloat32ToInt16(chunk);
-            const base64Data = arrayBufferToBase64(pcmData16.buffer);
-
-            await ipcRenderer.invoke('send-audio-content', {
-                data: base64Data,
-                mimeType: 'audio/pcm;rate=24000',
-            });
-        }
-    };
-
-    source.connect(audioProcessor);
-    audioProcessor.connect(audioContext.destination);
-}
-
-async function captureScreenshot(imageQuality = 'medium', isManual = false) {
-    console.log(`Capturing ${isManual ? 'manual' : 'automated'} screenshot...`);
-
-    // Check rate limiting for automated screenshots only
-    if (!isManual && tokenTracker.shouldThrottle()) {
-        console.log('⚠️ Automated screenshot skipped due to rate limiting');
-        return;
-    }
-
-    try {
-        // Request screenshot from main process
-        const result = await ipcRenderer.invoke('capture-screenshot', {
-            quality: imageQuality,
-        });
-
-        if (result.success && result.base64) {
-            // Store the latest screenshot
-            lastScreenshotBase64 = result.base64;
-
-            if (sendResult.success) {
-                // Track image tokens after successful send
-                const imageTokens = tokenTracker.calculateImageTokens(result.width || 1920, result.height || 1080);
-                tokenTracker.addTokens(imageTokens, 'image');
-                console.log(`📊 Image sent successfully - ${imageTokens} tokens used (${result.width}x${result.height})`);
-            } else {
-                console.error('Failed to send image:', sendResult.error);
-            }
-        } else {
-            console.error('Failed to capture screenshot:', result.error);
-        }
-    } catch (error) {
-        console.error('Error capturing screenshot:', error);
-    }
-}
-
-async function captureManualScreenshot(imageQuality = null) {
-    console.log('Manual screenshot triggered');
-    const quality = imageQuality || currentImageQuality;
-    await captureScreenshot(quality, true);
-}
-
-// Expose functions to global scope for external access
-window.captureManualScreenshot = captureManualScreenshot;
 
 function stopCapture() {
     if (screenshotInterval) {
@@ -706,76 +659,25 @@ function stopCapture() {
     }
 }
 
-async function getCurrentScreenshot() {
-    try {
-        // First try to get a fresh screenshot from main process
-        const result = await ipcRenderer.invoke('get-current-screenshot');
-
-        if (result.success && result.base64) {
-            console.log('📸 Got fresh screenshot from main process');
-            return result.base64;
-        }
-
-        // If no screenshot available, capture one now
-        console.log('📸 No screenshot available, capturing new one');
-        const captureResult = await ipcRenderer.invoke('capture-screenshot', {
-            quality: currentImageQuality,
-        });
-
-        if (captureResult.success && captureResult.base64) {
-            lastScreenshotBase64 = captureResult.base64;
-            return captureResult.base64;
-        }
-
-        // Fallback to last stored screenshot
-        if (lastScreenshotBase64) {
-            console.log('📸 Using cached screenshot');
-            return lastScreenshotBase64;
-        }
-
-        throw new Error('Failed to get screenshot');
-    } catch (error) {
-        console.error('Error getting current screenshot:', error);
-        return null;
-    }
-}
-
-function formatRealtimeConversationHistory() {
-    if (realtimeConversationHistory.length === 0) return 'No conversation history available.';
-
-    return realtimeConversationHistory.slice(-30).join('\n');
-}
-
-window.pickleGlass = {
-    initializeopenai,
+// ---------------------------
+// Exports & global registration
+// ---------------------------
+module.exports = {
     startCapture,
     stopCapture,
-    isLinux: isLinux,
-    isMacOS: isMacOS,
-    e: pickleGlassElement,
+    captureManualScreenshot,
+    getCurrentScreenshot,
+    isLinux,
+    isMacOS,
 };
 
-// -------------------------------------------------------
-// 🔔 React to session state changes from the main process
-// When the session ends (isActive === false), ensure we stop
-// all local capture pipelines (mic, screen, etc.).
-// -------------------------------------------------------
-ipcRenderer.on('session-state-changed', (_event, { isActive }) => {
-    if (!isActive) {
-        console.log('[Renderer] Session ended – stopping local capture');
-        stopCapture();
-    } else {
-        console.log('[Renderer] New session started – clearing in-memory history and summaries');
-
-        // Reset live conversation & analysis caches
-        realtimeConversationHistory = [];
-
-        const blankData = {
-            summary: [],
-            topic: { header: '', bullets: [] },
-            actions: [],
-            followUps: [],
-        };
-        window.pickleGlass.setStructuredData(blankData);
-    }
-});
+// Expose functions to global scope for external access (exact from renderer.js)
+if (typeof window !== 'undefined') {
+    window.captureManualScreenshot = captureManualScreenshot;
+    window.listenCapture = module.exports;
+    window.pickleGlass = window.pickleGlass || {};
+    window.pickleGlass.startCapture = startCapture;
+    window.pickleGlass.stopCapture = stopCapture;
+    window.pickleGlass.captureManualScreenshot = captureManualScreenshot;
+    window.pickleGlass.getCurrentScreenshot = getCurrentScreenshot;
+}
