+++ src/features/listen/stt/repositories/index.js
@@ -1,0 +1,5 @@
+const sttRepository = require('./sqlite.repository');
+
+module.exports = {
+    ...sttRepository,
+};
