+++ src/features/listen/summary/repositories/index.js
@@ -1,0 +1,5 @@
+const summaryRepository = require('./sqlite.repository');
+
+module.exports = {
+    ...summaryRepository,
+};
