+++ src/features/listen/summary/summaryService.js
@@ -1,0 +1,357 @@
+const { BrowserWindow } = require('electron');
+const { getSystemPrompt } = require('../../../common/prompts/promptBuilder.js');
+const { makeChatCompletionWithPortkey } = require('../../../common/services/aiProviderService.js');
+const authService = require('../../../common/services/authService');
+const sessionRepository = require('../../../common/repositories/session');
+const summaryRepository = require('./repositories');
+const { getStoredApiKey, getStoredProvider } = require('../../../electron/windowManager');
+
+class SummaryService {
+    constructor() {
+        this.previousAnalysisResult = null;
+        this.analysisHistory = [];
+        this.conversationHistory = [];
+        this.currentSessionId = null;
+        
+        // Callbacks
+        this.onAnalysisComplete = null;
+        this.onStatusUpdate = null;
+    }
+
+    setCallbacks({ onAnalysisComplete, onStatusUpdate }) {
+        this.onAnalysisComplete = onAnalysisComplete;
+        this.onStatusUpdate = onStatusUpdate;
+    }
+
+    setSessionId(sessionId) {
+        this.currentSessionId = sessionId;
+    }
+
+    async getApiKey() {
+        const storedKey = await getStoredApiKey();
+        if (storedKey) {
+            console.log('[SummaryService] Using stored API key');
+            return storedKey;
+        }
+
+        const envKey = process.env.OPENAI_API_KEY;
+        if (envKey) {
+            console.log('[SummaryService] Using environment API key');
+            return envKey;
+        }
+
+        console.error('[SummaryService] No API key found in storage or environment');
+        return null;
+    }
+
+    sendToRenderer(channel, data) {
+        BrowserWindow.getAllWindows().forEach(win => {
+            if (!win.isDestroyed()) {
+                win.webContents.send(channel, data);
+            }
+        });
+    }
+
+    addConversationTurn(speaker, text) {
+        const conversationText = `${speaker.toLowerCase()}: ${text.trim()}`;
+        this.conversationHistory.push(conversationText);
+        console.log(`💬 Added conversation text: ${conversationText}`);
+        console.log(`📈 Total conversation history: ${this.conversationHistory.length} texts`);
+
+        // Trigger analysis if needed
+        this.triggerAnalysisIfNeeded();
+    }
+
+    getConversationHistory() {
+        return this.conversationHistory;
+    }
+
+    resetConversationHistory() {
+        this.conversationHistory = [];
+        this.previousAnalysisResult = null;
+        this.analysisHistory = [];
+        console.log('🔄 Conversation history and analysis state reset');
+    }
+
+    /**
+     * Converts conversation history into text to include in the prompt.
+     * @param {Array<string>} conversationTexts - Array of conversation texts ["me: ~~~", "them: ~~~", ...]
+     * @param {number} maxTurns - Maximum number of recent turns to include
+     * @returns {string} - Formatted conversation string for the prompt
+     */
+    formatConversationForPrompt(conversationTexts, maxTurns = 30) {
+        if (conversationTexts.length === 0) return '';
+        return conversationTexts.slice(-maxTurns).join('\n');
+    }
+
+    async makeOutlineAndRequests(conversationTexts, maxTurns = 30) {
+        console.log(`🔍 makeOutlineAndRequests called - conversationTexts: ${conversationTexts.length}`);
+
+        if (conversationTexts.length === 0) {
+            console.log('⚠️ No conversation texts available for analysis');
+            return null;
+        }
+
+        const recentConversation = this.formatConversationForPrompt(conversationTexts, maxTurns);
+
+        // 이전 분석 결과를 프롬프트에 포함
+        let contextualPrompt = '';
+        if (this.previousAnalysisResult) {
+            contextualPrompt = `
+Previous Analysis Context:
+- Main Topic: ${this.previousAnalysisResult.topic.header}
+- Key Points: ${this.previousAnalysisResult.summary.slice(0, 3).join(', ')}
+- Last Actions: ${this.previousAnalysisResult.actions.slice(0, 2).join(', ')}
+
+Please build upon this context while analyzing the new conversation segments.
+`;
+        }
+
+        const basePrompt = getSystemPrompt('pickle_glass_analysis', '', false);
+        const systemPrompt = basePrompt.replace('{{CONVERSATION_HISTORY}}', recentConversation);
+
+        try {
+            if (this.currentSessionId) {
+                await sessionRepository.touch(this.currentSessionId);
+            }
+            
+            const messages = [
+                {
+                    role: 'system',
+                    content: systemPrompt,
+                },
+                {
+                    role: 'user',
+                    content: `${contextualPrompt}
+
+Analyze the conversation and provide a structured summary. Format your response as follows:
+
+**Summary Overview**
+- Main discussion point with context
+
+**Key Topic: [Topic Name]**
+- First key insight
+- Second key insight
+- Third key insight
+
+**Extended Explanation**
+Provide 2-3 sentences explaining the context and implications.
+
+**Suggested Questions**
+1. First follow-up question?
+2. Second follow-up question?
+3. Third follow-up question?
+
+Keep all points concise and build upon previous analysis if provided.`,
+                },
+            ];
+
+            console.log('🤖 Sending analysis request to AI...');
+
+            const API_KEY = await this.getApiKey();
+            if (!API_KEY) {
+                throw new Error('No API key available');
+            }
+            
+            const provider = getStoredProvider ? await getStoredProvider() : 'openai';
+            const loggedIn = authService.getCurrentUser().isLoggedIn;
+            const usePortkey = loggedIn && provider === 'openai';
+            
+            console.log(`[SummaryService] provider: ${provider}, usePortkey: ${usePortkey}`);
+
+            const completion = await makeChatCompletionWithPortkey({
+                apiKey: API_KEY,
+                provider: provider,
+                messages: messages,
+                temperature: 0.7,
+                maxTokens: 1024,
+                model: provider === 'openai' ? 'gpt-4.1' : 'gemini-2.5-flash',
+                usePortkey: usePortkey,
+                portkeyVirtualKey: usePortkey ? API_KEY : null
+            });
+
+            const responseText = completion.content;
+            console.log(`✅ Analysis response received: ${responseText}`);
+            const structuredData = this.parseResponseText(responseText, this.previousAnalysisResult);
+
+            if (this.currentSessionId) {
+                summaryRepository.saveSummary({
+                    sessionId: this.currentSessionId,
+                    text: responseText,
+                    tldr: structuredData.summary.join('\n'),
+                    bullet_json: JSON.stringify(structuredData.topic.bullets),
+                    action_json: JSON.stringify(structuredData.actions),
+                    model: 'gpt-4.1'
+                }).catch(err => console.error('[DB] Failed to save summary:', err));
+            }
+
+            // 분석 결과 저장
+            this.previousAnalysisResult = structuredData;
+            this.analysisHistory.push({
+                timestamp: Date.now(),
+                data: structuredData,
+                conversationLength: conversationTexts.length,
+            });
+
+            // 히스토리 크기 제한 (최근 10개만 유지)
+            if (this.analysisHistory.length > 10) {
+                this.analysisHistory.shift();
+            }
+
+            return structuredData;
+        } catch (error) {
+            console.error('❌ Error during analysis generation:', error.message);
+            return this.previousAnalysisResult; // 에러 시 이전 결과 반환
+        }
+    }
+
+    parseResponseText(responseText, previousResult) {
+        const structuredData = {
+            summary: [],
+            topic: { header: '', bullets: [] },
+            actions: [],
+            followUps: ['✉️ Draft a follow-up email', '✅ Generate action items', '📝 Show summary'],
+        };
+
+        // 이전 결과가 있으면 기본값으로 사용
+        if (previousResult) {
+            structuredData.topic.header = previousResult.topic.header;
+            structuredData.summary = [...previousResult.summary];
+        }
+
+        try {
+            const lines = responseText.split('\n');
+            let currentSection = '';
+            let isCapturingTopic = false;
+            let topicName = '';
+
+            for (const line of lines) {
+                const trimmedLine = line.trim();
+
+                // 섹션 헤더 감지
+                if (trimmedLine.startsWith('**Summary Overview**')) {
+                    currentSection = 'summary-overview';
+                    continue;
+                } else if (trimmedLine.startsWith('**Key Topic:')) {
+                    currentSection = 'topic';
+                    isCapturingTopic = true;
+                    topicName = trimmedLine.match(/\*\*Key Topic: (.+?)\*\*/)?.[1] || '';
+                    if (topicName) {
+                        structuredData.topic.header = topicName + ':';
+                    }
+                    continue;
+                } else if (trimmedLine.startsWith('**Extended Explanation**')) {
+                    currentSection = 'explanation';
+                    continue;
+                } else if (trimmedLine.startsWith('**Suggested Questions**')) {
+                    currentSection = 'questions';
+                    continue;
+                }
+
+                // 컨텐츠 파싱
+                if (trimmedLine.startsWith('-') && currentSection === 'summary-overview') {
+                    const summaryPoint = trimmedLine.substring(1).trim();
+                    if (summaryPoint && !structuredData.summary.includes(summaryPoint)) {
+                        // 기존 summary 업데이트 (최대 5개 유지)
+                        structuredData.summary.unshift(summaryPoint);
+                        if (structuredData.summary.length > 5) {
+                            structuredData.summary.pop();
+                        }
+                    }
+                } else if (trimmedLine.startsWith('-') && currentSection === 'topic') {
+                    const bullet = trimmedLine.substring(1).trim();
+                    if (bullet && structuredData.topic.bullets.length < 3) {
+                        structuredData.topic.bullets.push(bullet);
+                    }
+                } else if (currentSection === 'explanation' && trimmedLine) {
+                    // explanation을 topic bullets에 추가 (문장 단위로)
+                    const sentences = trimmedLine
+                        .split(/\.\s+/)
+                        .filter(s => s.trim().length > 0)
+                        .map(s => s.trim() + (s.endsWith('.') ? '' : '.'));
+
+                    sentences.forEach(sentence => {
+                        if (structuredData.topic.bullets.length < 3 && !structuredData.topic.bullets.includes(sentence)) {
+                            structuredData.topic.bullets.push(sentence);
+                        }
+                    });
+                } else if (trimmedLine.match(/^\d+\./) && currentSection === 'questions') {
+                    const question = trimmedLine.replace(/^\d+\.\s*/, '').trim();
+                    if (question && question.includes('?')) {
+                        structuredData.actions.push(`❓ ${question}`);
+                    }
+                }
+            }
+
+            // 기본 액션 추가
+            const defaultActions = ['✨ What should I say next?', '💬 Suggest follow-up questions'];
+            defaultActions.forEach(action => {
+                if (!structuredData.actions.includes(action)) {
+                    structuredData.actions.push(action);
+                }
+            });
+
+            // 액션 개수 제한
+            structuredData.actions = structuredData.actions.slice(0, 5);
+
+            // 유효성 검증 및 이전 데이터 병합
+            if (structuredData.summary.length === 0 && previousResult) {
+                structuredData.summary = previousResult.summary;
+            }
+            if (structuredData.topic.bullets.length === 0 && previousResult) {
+                structuredData.topic.bullets = previousResult.topic.bullets;
+            }
+        } catch (error) {
+            console.error('❌ Error parsing response text:', error);
+            // 에러 시 이전 결과 반환
+            return (
+                previousResult || {
+                    summary: [],
+                    topic: { header: 'Analysis in progress', bullets: [] },
+                    actions: ['✨ What should I say next?', '💬 Suggest follow-up questions'],
+                    followUps: ['✉️ Draft a follow-up email', '✅ Generate action items', '📝 Show summary'],
+                }
+            );
+        }
+
+        console.log('📊 Final structured data:', JSON.stringify(structuredData, null, 2));
+        return structuredData;
+    }
+
+    /**
+     * Triggers analysis when conversation history reaches 5 texts.
+     */
+    async triggerAnalysisIfNeeded() {
+        if (this.conversationHistory.length >= 5 && this.conversationHistory.length % 5 === 0) {
+            console.log(`🚀 Triggering analysis (non-blocking) - ${this.conversationHistory.length} conversation texts accumulated`);
+
+            this.makeOutlineAndRequests(this.conversationHistory)
+                .then(data => {
+                    if (data) {
+                        console.log('📤 Sending structured data to renderer');
+                        this.sendToRenderer('update-structured-data', data);
+                        
+                        // Notify callback
+                        if (this.onAnalysisComplete) {
+                            this.onAnalysisComplete(data);
+                        }
+                    } else {
+                        console.log('❌ No analysis data returned from non-blocking call');
+                    }
+                })
+                .catch(error => {
+                    console.error('❌ Error in non-blocking analysis:', error);
+                });
+        }
+    }
+
+    getCurrentAnalysisData() {
+        return {
+            previousResult: this.previousAnalysisResult,
+            history: this.analysisHistory,
+            conversationLength: this.conversationHistory.length,
+        };
+    }
+}
+
+module.exports = SummaryService;
